pyproject.toml
[project]
name = "customShop"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ian", email = "ruhyun92@gmail.com"},
]
dependencies = [
    "django>=5.0.2",
    "djangorestframework>=3.14.0",
    "python-dotenv>=1.0.1",
    "pillow>=10.2.0",
    "djangorestframework-simplejwt>=5.3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
